Use file-loader for SVG files with ._load_by_url_.svg filename endings
Modifies WebPack configuration to use file-loader for SVG files which
names end with ._load_by_url_.svg, while using react-svg-loader for any
other .svg files.

This feature should not be used in any new code, it is provided just to
faciliate rapid move of legacy code to the new version of WebPack
config.